+++ src/library/types.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier: Apache-2.0
 pragma solidity ^0.8.18;
 
 library Types {
@@ -12,18 +12,18 @@ library Types {
     // FuturesTradeUploadData
 
     struct FuturesTradeUploadData {
-        uint batch_id;
-        uint count;
+        uint256 batch_id;
+        uint256 count;
         FuturesTradeUpload[] trades;
     }
 
     struct FuturesTradeUpload {
-        uint trade_id;
-        uint match_id;
+        uint256 trade_id;
+        uint256 match_id;
         address account_id;
         string symbol;
         bool side;
-        uint trade_qty;
+        uint256 trade_qty;
         // signature for validate signed by real user
         bytes signature;
     }
@@ -31,62 +31,62 @@ library Types {
     // EventUpload
 
     struct EventUpload {
-        uint batch_id;
-        uint count;
+        uint256 batch_id;
+        uint256 count;
         EventUploadData[] events;
     }
 
     struct EventUploadData {
-        uint event_id;
+        uint256 event_id;
         string biz_type;
-        uint biz_id;
+        uint256 biz_id;
         WithdrawData[] withdraws;
         Settlement[] settlements;
         Liquidation[] liquidations;
-        uint[] sequence;
+        uint256[] sequence;
     }
 
     struct WithdrawData {
         address account_id;
         bytes32 token;
-        uint withdraw_id;
+        uint256 withdraw_id;
         bool approval;
-        uint chain_id;  // target withdraw chain
+        uint256 chain_id; // target withdraw chain
     }
 
     struct Settlement {
         address account_id;
-        int settled_amount;
+        int256 settled_amount;
         bytes32 settled_asset;
-        uint insurance_transfer_amount;
+        uint256 insurance_transfer_amount;
         SettlementExecution[] settlement_executions;
     }
 
     struct Liquidation {
         address account_id;
-        int settled_amount;
+        int256 settled_amount;
         LiquidationTransfer[] liquidation_transfers;
-        uint timestamp;
+        uint256 timestamp;
         bytes32 liquidated_asset;
     }
 
     struct LiquidationTransfer {
-        uint liquidation_transfer_id;
+        uint256 liquidation_transfer_id;
         address liquidator_account_id;
         bytes32 list_symbol;
-        int position_qty_transfer;
-        int cost_position_transfer;
-        uint liquidator_fee;
-        uint insurance_fee;
-        uint mark_price;
-        int sum_unitary_fundings;
-        uint liquidation_fee;
+        int256 position_qty_transfer;
+        int256 cost_position_transfer;
+        uint256 liquidator_fee;
+        uint256 insurance_fee;
+        uint256 mark_price;
+        int256 sum_unitary_fundings;
+        uint256 liquidation_fee;
     }
 
     struct SettlementExecution {
         bytes32 symbol;
-        int sum_unitary_fundings;
-        uint mark_price;
-        int settled_amount;
+        int256 sum_unitary_fundings;
+        uint256 mark_price;
+        int256 settled_amount;
     }
 }
